Lots of changes Long ago i moved the BinUtils parsing code off to another library BinaryEx, so i've migrated the whole repo to use it. Upgrade to .net 7 for testing/development Upgrade nuget packages Fix several bugs, the roundtrip test now passes. removed all tests that were for the binstream code Implemented a faster ReadVlv function that has had several iterations over the years. Switch things to be able to inline over assembly boundries

+++ NoFrillSMF/Events/MetaEventParsing.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics;
 using System;
-using NoFrill.Common;
+using BinaryEx;
 using NoFrillSMF.Chunks;
 
 namespace NoFrillSMF.Events
@@ -13,7 +13,7 @@ namespace NoFrillSMF.Events
             metaEvent.Status = state.status;
             metaEvent.eventType = (EventType)data.ReadByte(ref offset);
             Debug.Assert(metaEvent.Status == 0xFF);
-            metaEvent.Size = data.ReadVlv(ref offset);
+            metaEvent.Size = (int)data.ReadVlv(ref offset);
         }
 
         public static void ComposeStatus<T>(this T metaEvent, byte[] data, ref int offset) where T : BaseTrackEvent
